requirements.txt

# uv pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple
git+https://github.com/huggingface/huggingface_hub
fastapi 
uvicorn[standard]
gradio 
pillow 
huggingface_hub 
gradio_client
python-dotenv
python-multipart
requests
